pyproject.toml
[project]
name = "financeresearchagent"
version = "0.1.0"
description = "AI-powered investment research and decision support agent"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "langchain>=1.0.0",
    "langgraph>=1.0.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.2.0",
    "yfinance>=0.2.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "jinja2>=3.1.0",
    "markdown>=3.5.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.135.1",
    "uvicorn>=0.41.0",
    "deepagents>=0.4.7",
    "build>=1.4.0",
    "pyproject-hooks>=1.2.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "ipython>=8.0.0",
]
api = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sse-starlette>=2.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
